Code Exercise 06 “Counting Cards”
The user inputs four players names. The computer then deals each play all the cards in a deck 2-A. The players’ cards are revealed and totaled up to determine who has the highest total. The players and their totaled are listed in 1st, 2nd, 3rd, 4th place order.
2-10 = Card value
J,Q,K = 12
A = 15