My Beer Application
This application is a beer menu for local breweries and restaurants. It requires a user to login or create an account. Once logged in they must select a location with each location having their own list of beers. The user can select to view the menu sorting the list by beer name, ABV, or by beer style. Inside the application a user can also order a beer which the data is printed in the form of a receipt to the establistments kitchen or staff as well as stored inside a database. In addition the user can add and/or remove their favorite beers from the entire list of beer (2000+) under their favorites tab. lastly user's can write reviews about each beer also giving it a rating 1-5.