Time Tracker Application
Time Tracker is an application that adds and manages life events for display. It also runs calulations showing the user how much, percentage, etc that spend on a certain activity every day, week, and month. The information is stored in a database and pushed and pulled from the application for use.